fix: stash
